+++ packages/analytics/dashboard-renderer/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.9](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@0.1.8...@kong-ui-public/dashboard-renderer@0.1.9) (2023-11-29)
+
+**Note:** Version bump only for package @kong-ui-public/dashboard-renderer
+
+
+
+
+
+## [0.1.8](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@0.1.7...@kong-ui-public/dashboard-renderer@0.1.8) (2023-11-28)
+
+**Note:** Version bump only for package @kong-ui-public/dashboard-renderer
+
+
+
+
+
+## [0.1.7](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@0.1.6...@kong-ui-public/dashboard-renderer@0.1.7) (2023-11-28)
+
+**Note:** Version bump only for package @kong-ui-public/dashboard-renderer
+
+
+
+
+
+## [0.1.6](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@0.1.5...@kong-ui-public/dashboard-renderer@0.1.6) (2023-11-27)
+
+**Note:** Version bump only for package @kong-ui-public/dashboard-renderer
+
+
+
+
+
+## [0.1.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@0.1.4...@kong-ui-public/dashboard-renderer@0.1.5) (2023-11-24)
+
+**Note:** Version bump only for package @kong-ui-public/dashboard-renderer
+
+
+
+
+
+## [0.1.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@0.1.3...@kong-ui-public/dashboard-renderer@0.1.4) (2023-11-22)
+
+**Note:** Version bump only for package @kong-ui-public/dashboard-renderer
+
+
+
+
+
 ## [0.1.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/dashboard-renderer@0.1.2...@kong-ui-public/dashboard-renderer@0.1.3) (2023-11-21)
 
 **Note:** Version bump only for package @kong-ui-public/dashboard-renderer