+++ packages/core/app-layout/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/app-layout@2.1.2...@kong-ui-public/app-layout@2.1.3) (2023-11-29)
+
+**Note:** Version bump only for package @kong-ui-public/app-layout
+
+
+
+
+
+## [2.1.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/app-layout@2.1.1...@kong-ui-public/app-layout@2.1.2) (2023-11-28)
+
+**Note:** Version bump only for package @kong-ui-public/app-layout
+
+
+
+
+
+## [2.1.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/app-layout@2.1.0...@kong-ui-public/app-layout@2.1.1) (2023-11-22)
+
+
+### Bug Fixes
+
+* **entities-vaults:** fix kradio card occurrences ([#937](https://github.com/Kong/public-ui-components/issues/937)) ([21827e9](https://github.com/Kong/public-ui-components/commit/21827e9592a962ec0e7a7de1d76ebd80c36dcd2f))
+
+
+
+
+
 # [2.1.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/app-layout@2.0.2...@kong-ui-public/app-layout@2.1.0) (2023-11-21)