+++ packages/core/core/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.9](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/core@1.4.8...@kong-ui-public/core@1.4.9) (2023-11-29)
+
+**Note:** Version bump only for package @kong-ui-public/core
+
+
+
+
+
+## [1.4.8](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/core@1.4.7...@kong-ui-public/core@1.4.8) (2023-11-28)
+
+**Note:** Version bump only for package @kong-ui-public/core
+
+
+
+
+
 ## [1.4.7](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/core@1.4.6...@kong-ui-public/core@1.4.7) (2023-11-21)
 
 **Note:** Version bump only for package @kong-ui-public/core